function/Function1/requirements.txt
adal==1.2.3
astroid==2.4.1
azure-common==1.1.25
azure-functions==1.2.1
azure-servicebus==0.50.2
backports.entry-points-selectable==1.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.2
dominate==2.5.1
feedgen==0.9.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.1
idna==2.9
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.13
msrestazure==0.6.3
oauthlib==3.1.0
pipenv==2021.5.29
platformdirs==2.3.0
psycopg2==2.8.5
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
python-http-client==3.2.7
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.3.0
shell==1.0.1
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
uamqp==1.2.7
urllib3==1.25.9
virtualenv==20.7.2
virtualenv-clone==0.5.7
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
